Split Duid class into multiple concern based class partials
